requirements.txt
selenium==2.53.0
#chromedriver-binary==73.0.3683.68.0
boto3==1.6.9
boto==2.48.0
botocore==1.9.9
pyyaml==3.12